Extract props to common interface
